feat(mobile): optimize bundle size, re-renders, and startup time
- Implement lazy loading for profile and settings screens (#127)
- Add React.memo to components to reduce unnecessary re-renders (#128)
- Add startup time measurement and performance tracking (#129)

Closes #127, #128, #129
