+++ mobile/components/screens/ProfileScreenContent.tsx
@@ -1,4 +1,6 @@
-import React, { useState, useMemo } from 'react';
+'use client';
+
+import React, { useState } from 'react';
 import {
   SafeAreaView,
   ScrollView,
@@ -6,28 +8,28 @@ import {
   Text,
   Pressable,
   StyleSheet,
-  Switch,
 } from 'react-native';
+import { useTranslation } from 'react-i18next';
 import { useRouter } from 'expo-router';
-import { Avatar } from '../../../components/ui/Avatar';
-import { Button } from '../../../components/ui/Button';
-import { DisconnectModal } from '../../../components/wallet/DisconnectModal';
-import { useAuthStore } from '../../../store/authStore';
+import { Avatar } from '../../components/ui/Avatar';
+import { Button } from '../../components/ui/Button';
+import { DisconnectModal } from '../../components/wallet/DisconnectModal';
+import { useAuthStore } from '../../store/authStore';
+import { shallow } from 'zustand/shallow';
 
 // Truncate wallet address
 function truncateAddress(address: string): string {
   return `${address.slice(0, 6)}...${address.slice(-4)}`;
 }
 
-const ProfileScreen = React.memo(() => {
+export default function ProfileScreenContent() {
+  const { t } = useTranslation();
   const router = useRouter();
-  const wallet = useAuthStore((s) => s.wallet);
-  const logout = useAuthStore((s) => s.logout);
+  const [wallet, logout] = useAuthStore((state) => [state.wallet, state.logout], shallow);
   const [disconnectModalVisible, setDisconnectModalVisible] = useState(false);
 
-  const displayName = 'EsuStellar User';
+  const displayName = wallet ? truncateAddress(wallet.publicKey) : t('home.defaultUser');
   const walletAddress = wallet?.publicKey || 'GABCD1234EFGH5678IJKL9012MNOP';
-  const truncatedAddress = useMemo(() => truncateAddress(walletAddress), [walletAddress]);
 
   const handleDisconnect = () => {
     logout();
@@ -41,13 +43,13 @@ const ProfileScreen = React.memo(() => {
         <View style={styles.header}>
           <Avatar name={displayName} size="lg" />
           <Text style={styles.displayName}>{displayName}</Text>
-          <Text style={styles.walletAddress}>{truncatedAddress}</Text>
+          <Text style={styles.walletAddress}>{truncateAddress(walletAddress)}</Text>
         </View>
 
         {/* Action Button */}
         <View style={styles.section}>
           <Button variant="outline" size="lg" onPress={() => router.push('/profile/edit')}>
-            Edit Profile
+            {t('profile.editProfile')}
           </Button>
         </View>
 
@@ -55,10 +57,10 @@ const ProfileScreen = React.memo(() => {
         <View style={styles.section}>
           <Pressable
             style={styles.settingsRow}
-            onPress={() => router.push('/profile/settings')}
+            onPress={() => router.push('/settings')}
             accessibilityRole="button"
           >
-            <Text style={styles.settingsLabel}>Settings</Text>
+            <Text style={styles.settingsLabel}>{t('profile.settings')}</Text>
             <Text style={styles.chevron}>›</Text>
           </Pressable>
 
@@ -69,8 +71,8 @@ const ProfileScreen = React.memo(() => {
             onPress={() => setDisconnectModalVisible(true)}
             accessibilityRole="button"
           >
-            <Text style={[styles.settingsLabel, { color: '#EF4444' }]}>
-              Disconnect Wallet
+            <Text style={[styles.settingsLabel, { color: '#EF4444' }]}> 
+              {t('profile.disconnectWallet')}
             </Text>
             <Text style={[styles.chevron, { color: '#EF4444' }]}>›</Text>
           </Pressable>
@@ -86,9 +88,9 @@ const ProfileScreen = React.memo(() => {
       />
     </SafeAreaView>
   );
-});
+}
 
-export default ProfileScreen;
+const styles = StyleSheet.create({
   container: {
     flex: 1,
     backgroundColor: '#0F172A',
